Different links for different artists, commas, hover effects and start of read me
