+++ code/src/Components/Header.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+// The header as it own Component
+export const Header = () => {
+    return ( <h1 className="header" tabIndex="0"> New albums & singles </h1>)
+    };
