+++ code/src/Components/Icons.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+// Inserting the icons
+export const Icons = () => (
+  <div className="icons-container">
+    <img className="heartIcon" src="../icons/heart.svg" alt="Heart icon" tabIndex="0" />
+    <img className="playIcon" src="../icons/play.svg" alt="Play icon" tabIndex="0" />
+    <img className="dotsIcon" src="../icons/dots.svg" alt="Three dots" tabIndex="0" />
+  </div>
+);
+
