--- code/src/Header.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const Header = () => {
-    return ( < h1 className = "header" > New albums & singles < /h1>)
-    };
