--- code/src/Icons.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-export const Icons = () => (
-  <div className="icons-container">
-    <img className="heartIcon" src="../icons/heart.svg" alt="Heart icon" />
-    <img className="playIcon" src="../icons/play.svg" alt="Play icon" />
-    <img className="dotsIcon" src="../icons/dots.svg" alt="Three dots" />
-  </div>
-);
-
